% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Account.R
\name{account_attrs}
\alias{account_attrs}
\title{Account Attributes}
\usage{
account_attrs(attribute, ...)
}
\arguments{
\item{attribute}{Optionally a character string specifying one or more of: \dQuote{supported-platforms}, \dQuote{vpc-max-security-groups-per-interface}, \dQuote{max-elastic-ips}, \dQuote{max-instances}, \dQuote{vpc-max-elastic-ips}, \dQuote{default-vpc}. If missing, all are returned.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Retrieve Account Attributes
}
\examples{
\dontrun{
account_attrs()
}
}
