% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.R
\name{allocate_ip}
\alias{allocate_ip}
\alias{release_ip}
\title{Allocate/Release IP Addresses}
\usage{
allocate_ip(domain = c("vpc", "standard"), ...)

release_ip(ip, ...)
}
\arguments{
\item{domain}{Optionally, a character string specifying \dQuote{vpc} or \dQuote{standard}.}

\item{...}{Additional arguments passed to [ec2HTTP()].}

\item{ip}{An object of class \dQuote{ec2_ip} or an allocationId (for a VPC IP) or publicIp (for a standard IP).}
}
\value{
For `allocate_ip`, a list containing the IP address. For `release_ip`, a logical.
}
\description{
Allocate or release VPC or standard IP Address
}
\details{
This function is used to allocate IP addresses either to EC2 classic (if it is available on your account; see [account_attrs()]) or on a Virtual Private Cloud (VPC). The default for new AWS EC2 accounts is only to be able to create VPC configurations. Due to limitations in the IPv4 universe, users are typically restricted to 5 IP addresses, which can dynamically be allocated to instances via a VPC. Use [associate_ip()]/[disassociate_ip()] to link an IP address to a specific instance.
}
\examples{
\dontrun{
# create a classic/"standard" IP address
a1 <- allocate_ip("standard")
release_ip(a1$publicIp)

# create a VPC IP address
a2 <- allocate_ip("vpc")
release_ip(a2$allocationId)
}
}
\references{
<http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_AllocateAddress.html>
<http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_ReleaseAddress.html>
}
\seealso{
[associate_ip()], [describe_ips()], [release_ip()], [make_ip_vpc()]/[make_ip_classic()]
}
