% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.r
\name{associate_ip}
\alias{associate_ip}
\alias{disassociate_ip}
\title{(Dis)Associate IP}
\usage{
associate_ip(instance, allocation, public, private, netinterface, allow, ...)

disassociate_ip(allocation, public, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}.}

\item{allocation}{\dots}

\item{public}{\dots}

\item{private}{\dots}

\item{netinterface}{\dots}

\item{allow}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list.
}
\description{
Associate/Disassociate IP with Instance
}

