% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS.r
\name{attach_volume}
\alias{attach_volume}
\title{Attach volume}
\usage{
attach_volume(device, instance, volume, ...)
}
\value{
A list containing the results of the API call
}
\description{
Attach a volume to an instance on the given device
}

