% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.R
\name{make_ip_vpc}
\alias{make_ip_vpc}
\alias{make_ip_classic}
\title{IP conversion}
\usage{
make_ip_vpc(ip, ...)

make_ip_classic(ip, ...)
}
\arguments{
\item{ip}{An object of class \dQuote{ec2_ip} or an allocationId (for a VPC IP) or publicIp (for a standard IP).}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list.
}
\description{
Convert IP between Classic/EC2
}
\details{
IP addresses can only be used within EC2 classic or a Virtual Private Cloud. These functions allow you to convert an IP address between the two systems.
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_RestoreAddressToClassic.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_RestoreAddressToClassic.html}
}
