% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Images.r
\name{copy_image}
\alias{copy_image}
\title{Copy AMI}
\usage{
copy_image(name, image, description, region = getOption("AWS_DEFAULT_REGION",
  "us-east-1"), token, ...)
}
\arguments{
\item{name}{A character string specifying a name for the new AMI}

\item{image}{A character string containing an Amazon Machine Image (AMI) image ID or an object of class \dQuote{ec2_image}.}

\item{region}{A character string specifying the AWS region to create the AMI in. The default is \dQuote{us-east-1}.}

\item{token}{A (max 64-character) string containing a unique user-defined token. This can be used to ensure idempotency of requests in the event of a user-side failure. See \url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html} for details.}
}
\value{
A list.
}
\description{
Create a copy of an AMI

A character string containing a description of the AMI.
}
\examples{
\dontrun{
copy_image("example-rstudio-ami", "ami-7f9dc615", "This is a description")

}
}

