% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS.r
\name{create_snapshot}
\alias{create_snapshot}
\title{Create snapshot}
\usage{
create_snapshot(volume, description, ...)
}
\arguments{
\item{volume}{the volume-id of the volume to be snapshotted}

\item{description}{a description (metadata) for the volume}
}
\value{
A list containing the results of the API call
}
\description{
Create a snapshot of an EBS volume
}

