% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS.r
\name{delete_snapshot}
\alias{delete_snapshot}
\title{Delete snapshot}
\usage{
delete_snapshot(snapshot, ...)
}
\arguments{
\item{snapshot}{the snapshot-id to be deleted}
}
\value{
A list containing the results of the API call
}
\description{
Delete a given shapshot
}

