% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS.r
\name{delete_volume}
\alias{delete_volume}
\title{Delete volume}
\usage{
delete_volume(volume, ...)
}
\arguments{
\item{volume}{the volume-id to be deleted}
}
\value{
A list containing the results of the API call
}
\description{
Delete a given volume
}

