% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_account_attributes.R
\name{describe_account_attributes}
\alias{describe_account_attributes}
\title{Describe Account Attributes}
\usage{
describe_account_attributes(
  attribute_names = c("supported_platforms", "default_vpc",
    "vpc_max_security_groups_per_interface", "max_elastic_ips", "vpc_max_elastic_ips"),
  ...
)
}
\arguments{
\item{attribute_names}{A character vector specifying the attributes to
include. One or more of \code{supported_platforms}, \code{default_vpc},
\code{vpc_max_security_groups_per_interface}, \code{max_elastic_ips}, or
\code{vpc_max_elastic_ips}.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}. Note that
the \code{dry_run} parameter can change the return of this function.}
}
\value{
A list of the requested attributes.
}
\description{
Retrieve Account Attributes
}
\examples{
\dontrun{
describe_account_attributes()
}
}
\references{
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeAccountAttributes.html}{API
Reference: DescribeAccountAttributes}
}
