% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gateways.R
\name{describe_gateways}
\alias{describe_gateways}
\title{describe_gateways}
\usage{
describe_gateways(gateway, filter, ...)
}
\arguments{
\item{gateway}{A character string specifying an Internet Gateway ID, or an object of class \dQuote{ec2_internet_gateway}.}

\item{filter}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list.
}
\description{
Describe Network Gateways
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeInternetGateways.html}
}
\seealso{
\code{\link{create_gateway}}, \code{\link{attach_gateway}}
}
\keyword{security}
