% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Images.R
\name{describe_images}
\alias{describe_images}
\title{Describe AMI(s)}
\usage{
describe_images(image, filter, availableto, owner, ...)
}
\arguments{
\item{image}{A character string containing an Amazon Machine Image (AMI) image ID or an object of class \dQuote{ec2_image}.}

\item{filter}{\dots}

\item{availableto}{Scopes the images by users with explicit launch
permissions. Specify an AWS account ID, \code{"self"} (the sender of the
request), or `"all"`` (public AMIs).}

\item{owner}{Filters the images by the owner. Specify an AWS account ID,
\code{"self"} (owner is the sender of the request), or an AWS owner alias (valid
values are \code{"amazon"} | \code{"aws-marketplace"} | \code{"microsoft"}). Omitting this
option returns all images for which you have launch permissions, regardless
of ownership.}

\item{...}{Additional arguments passed to [ec2HTTP()].}
}
\value{
A list
}
\description{
Search/Describe AMI(s)
}
\examples{
\dontrun{
# RStudio AMIs from: http://www.louisaslett.com/RStudio_AMI/
describe_images("ami-7f9dc615")

# Amazon Linux AMI from: http://aws.amazon.com/amazon-linux-ami/
describe_images("ami-08111162")
}
}
\references{
https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeImages.html
}
\keyword{images}
