% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instances.r
\name{describe_instances}
\alias{describe_instances}
\alias{instance_status}
\title{EC2 Instance Status}
\usage{
describe_instances(instance, filter, n, token, ...)

instance_status(instance, filter, runningonly, n, token, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}.}

\item{filter}{\dots}

\item{n}{\dots}

\item{token}{A (max 64-character) string containing a unique user-defined token. This can be used to ensure idempotency of requests in the event of a user-side failure. See \url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html} for details.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{runningonly}{\dots}
}
\value{
A list
}
\description{
Describe or check status of an EC2 instance
}
\examples{
\dontrun{
i <- run_instance()
describe_instances(i[[1]])
instance_status(i[[1]])

stop_instances(i[[1]])
terminate_instances(i[[1]])
}
}

