% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.r
\name{describe_ip}
\alias{describe_ip}
\title{Describe IP}
\usage{
describe_ip(ip, filter, ...)
}
\arguments{
\item{ip}{An object of class \dQuote{ec2_ip} or an allocationId (for a VPC IP) or publicIp (for a standard IP).}

\item{filter}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Get IP information
}

