% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.R
\name{describe_ips}
\alias{describe_ips}
\title{Describe IP(s)}
\usage{
describe_ips(ip, filter, ...)
}
\arguments{
\item{ip}{An object of class \dQuote{ec2_ip} or an allocationId (for a VPC IP) or publicIp (for a standard IP).}

\item{filter}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Get information about one or more IP addresses.
}
