% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_regions.R
\name{describe_regions}
\alias{describe_regions}
\title{Describe Regions}
\usage{
describe_regions(region = NULL, filter = NULL, ...)
}
\arguments{
\item{region}{Optionally a character string specifying the name of one or more AWS regions. If missing, details about all are returned.}

\item{filter}{\dots}

\item{...}{Additional arguments passed to [ec2HTTP()].}
}
\value{
A list
}
\description{
Retrieve Region Details
}
\examples{
\dontrun{
decribe_regions()
}
}
\seealso{
\code{\link[=describe_zones]{describe_zones()}}
}
