% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS.r
\name{describe_snapshots}
\alias{describe_snapshots}
\title{Describe snapshots}
\usage{
describe_snapshots(filter, n, page, ...)
}
\arguments{
\item{filter}{\dots}

\item{n}{\dots}

\item{page}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list of lists, each list containing API returned data on matching snapshots
}
\description{
Search/Describe snapshots
}

