% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS.r
\name{describe_volumes}
\alias{describe_volumes}
\title{Describe volumes}
\usage{
describe_volumes(filter, n, page, ...)
}
\arguments{
\item{filter}{\dots}

\item{n}{\dots}

\item{page}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list of lists, each list containing API returned data on matching snapshots
}
\description{
Search/Describe volumes
}

