% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.R
\name{describe_vpcs}
\alias{describe_vpcs}
\title{Describe VPC(s)}
\usage{
describe_vpcs(vpc, filter, ...)
}
\arguments{
\item{vpc}{A character string specifying a VPC Allocation ID or an object of class \dQuote{ec2_ip}.}

\item{filter}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list.
}
\description{
List/describe VPC(s)
}
\examples{
\dontrun{
describe_vpcs()
}
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVpcs.html}
}
\seealso{
\code{\link{create_vpc}}, \code{\link{get_vpc_attr}}
}
\keyword{security}
