% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Account.R
\name{describe_zones}
\alias{describe_zones}
\title{Describe Zones}
\usage{
describe_zones(zone, filter, ...)
}
\arguments{
\item{zone}{Optionally a character string specifying the name of one or more AWS zones. If missing, details about all are returned.}

\item{filter}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Retrieve Zone Details
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeAvailabilityZones.html}
}
\seealso{
\code{\link{describe_regions}}
}
