% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS.r
\name{detach_volume}
\alias{detach_volume}
\title{Detach volume}
\usage{
detach_volume(device, instance, volume, force, ...)
}
\value{
A list containing the results of the API call
}
\description{
Detach a volume to an instance on the given device
}

