% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebs_volumes.R
\name{create_volume}
\alias{create_volume}
\alias{modify_volume}
\alias{delete_volume}
\alias{attach_volume}
\alias{detach_volume}
\alias{enable_volume_io}
\alias{get_volume_attr}
\alias{set_volume_attr}
\alias{describe_volumes}
\alias{volume_status}
\title{Elastic Block Store Volumes}
\usage{
create_volume(size = NULL, type = NULL, iops = NULL, snapshot = NULL,
  ...)

modify_volume(volume, size = NULL, type = NULL, iops = NULL, ...)

delete_volume(volume, ...)

attach_volume(volume, instance, device, ...)

detach_volume(volume, instance, device = NULL, force = FALSE, ...)

enable_volume_io(volume, ...)

get_volume_attr(volume, attribute = c("autoEnableIO", "productCodes"), ...)

set_volume_attr(volume, auto_enable, ...)

describe_volumes(volume = NULL, filter = NULL, ...)

volume_status(volume = NULL, filter = NULL, ...)
}
\arguments{
\item{size}{Size of the volume in gigabytes. See API documentation for allowed values.}

\item{type}{The volume type. Defaults to "Standard".}

\item{iops}{The number of I/O operations per second (IOPS) to provision for the volume, ranging from 100 to 32000.}

\item{snapshot}{A character string containing an EBS Volume Snapshot snapshot ID or an object of class \dQuote{ec2_ebs_snapshot}.}

\item{...}{Additional arguments passed to [ec2HTTP()].}

\item{volume}{A character string containing an EBS Volume volume ID, or an object of class \dQuote{ec2_ebs_volume}.}

\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}. Or, actions can be taken on multiple instances in one call by passing a character vector of instance IDs, or a list of objects of class \dQuote{ec2_instance}.}

\item{device}{\dots}

\item{force}{\dots}

\item{attribute}{\dots}

\item{auto_enable}{\dots}

\item{filter}{\dots}
}
\value{
\dots
}
\description{
Manage Elastic Block Store (EBS) volumes for EC2
}
\references{
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreateVolume.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DeleteVolume.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_ModifyVolume.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_AttachVolume.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DetachVolume.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVolumes.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVolumeStatus.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVolumeAttribute.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_ModifyVolumeAttribute.html}
}
\seealso{
\code{\link[=run_instances]{run_instances()}}, \code{\link[=create_snapshot]{create_snapshot()}}
}
