% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebs_snapshots.R
\name{create_snapshot}
\alias{create_snapshot}
\alias{delete_snapshot}
\title{Elastic Block Store Snapshots}
\usage{
create_snapshot(volume, description = NULL, ...)

delete_snapshot(snapshot, ...)
}
\arguments{
\item{volume}{A character string containing an EBS Volume volume ID, or an object of class \dQuote{ec2_ebs_volume}.}

\item{description}{\dots}

\item{...}{Additional arguments passed to [ec2HTTP()].}

\item{snapshot}{A character string containing an EBS Volume Snapshot snapshot ID or an object of class \dQuote{ec2_ebs_snapshot}.}
}
\value{
\dots
}
\description{
Manage Elastic Block Store (EBS) volumes for EC2
}
\references{
<https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreateVolume.html>
<https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DeleteVolume.html>
<https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_ModifyVolume.html>
<https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_AttachVolume.html>
<https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DetachVolume.html>
<https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVolumes.html>
<https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVolumeStatus.html>
<https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVolumeAttribute.html>
<https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_ModifyVolumeAttribute.html>
}
\seealso{
[run_instances()], [create_volume()]
}
