% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gateways.R
\name{create_gateway}
\alias{create_gateway}
\alias{delete_gateway}
\title{Create/Delete Gateways}
\usage{
create_gateway(...)

delete_gateway(gateway, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{gateway}{A character string specifying an Internet Gateway ID, or an object of class \dQuote{ec2_internet_gateway}.}
}
\description{
Create and delete Network Gateways for a VPC
}
\references{
\url{http://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreateInternetGateway.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DeleteInternetGateway.html}
}
\seealso{
\code{\link{create_gateway}}, \code{\link{describe_gateways}}
}
\keyword{security}
