% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instances.r
\name{get_console_output}
\alias{get_console_output}
\title{Get EC2 Instance Console Output}
\usage{
get_console_output(instance, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A character string
}
\description{
Retrieve console output for an EC2 instance as a character string
}
\examples{
\dontrun{
i <- run_instance()
cat(get_console_output(), "\\n")
stop_instances(i)
terminate_instances(i)
}
}

