% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Images.r
\name{get_image_attr}
\alias{get_image_attr}
\title{Get AMI Attribute}
\usage{
get_image_attr(image, attribute, ...)
}
\arguments{
\item{image}{A character string containing an Amazon Machine Image (AMI) image ID or an object of class \dQuote{ec2_image}}

\item{attribute}{A character string specifying one of: \dQuote{description}, \dQuote{kernel}, \dQuote{ramdisk}, \dQuote{launchPermission}, \dQuote{productCodes}, \dQuote{blockDeviceMapping}, \dQuote{sriovNetSupport}}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Get AMI Attribute
}
\examples{
\dontrun{
# RStudio AMIs from: http://www.louisaslett.com/RStudio_AMI/
get_image_attr("ami-7f9dc615", "description")
}
}

