% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_parameter.R
\name{get_parameter}
\alias{get_parameter}
\title{gets a parameter from the parameter store}
\usage{
get_parameter(name, with_decryption = TRUE, ...)
}
\arguments{
\item{name}{The parameter name}

\item{with_decryption}{Specifies if the parameter should be decripted. Defaults to \code{TRUE}}

\item{...}{Additional arguments passed to \code{ec2HTTP}}
}
\value{
The value of the parameter
}
\description{
gets a parameter from the parameter store
}
