% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keypair.R
\name{describe_keypairs}
\alias{describe_keypairs}
\alias{create_keypair}
\alias{delete_keypair}
\alias{import_keypair}
\title{EC2 Keypairs}
\usage{
describe_keypairs(keypair, filter, ...)

create_keypair(keypair, path, ...)

delete_keypair(keypair, ...)

import_keypair(keypair, publickey, ...)
}
\arguments{
\item{keypair}{A character string (max 255 character) containing the name of a keypair, or an object of class \dQuote{ec2_keypair}.}

\item{filter}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{path}{A character string specifying a filepath (or filename) to use as a .pem file for the keypair.}

\item{publickey}{\dots}
}
\value{
A list of class \dQuote{ec2_keypair} or list of such objects.
}
\description{
Get, create, delete, and import EC2 keypairs
}
\examples{
\dontrun{
k <- create_keypair("test_keypair")
describe_keypairs()
delete_keypair(k)
}
}
\keyword{security}
