% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_instances.R
\name{monitor_instances}
\alias{monitor_instances}
\alias{unmonitor_instances}
\title{EC2 Instance Monitoring}
\usage{
monitor_instances(instance = NULL, ...)

unmonitor_instances(instance = NULL, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}. Or, actions can be taken on multiple instances in one call by passing a character vector of instance IDs, or a list of objects of class \dQuote{ec2_instance}.}

\item{...}{Additional arguments passed to [ec2HTTP()].}
}
\value{
A list.
}
\description{
Set EC2 instance monitoring in CloudWatch on or off.
}
\examples{
\dontrun{
i <- run_instances()
monitor_instances(i[[1]])
unmonitor_instances(i[[1]])

stop_instances(i[[1]])
terminate_instances(i[[1]])
}
}
\references{
<http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-cloudwatch.html>
<http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_MonitorInstances.html>
<http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_UnmonitorInstances.html>
}
\seealso{
[describe_instances()]
}
\keyword{instances}
