% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlacementGroups.R
\name{create_placement}
\alias{create_placement}
\alias{delete_placement}
\alias{describe_placements}
\title{Placement Groups}
\usage{
create_placement(group, strategy = "cluster", ...)

delete_placement(group, ...)

describe_placements(group, filter, ...)
}
\arguments{
\item{group}{A character string specifying a placement group.}

\item{strategy}{A character string specifying \dQuote{cluster}. No other values are currently allowed.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{filter}{\dots}
}
\value{
For \code{describe_placements}, a list of objects of class \dQuote{ec2_placement_group}. Otherwise, a logical.
}
\description{
EC2 Cluster \dQuote{Placement Groups}
}
\examples{
\dontrun{
pg <- create_placement("examplepg")
describe_placements()
delete_placement("examplepg")
}
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/placement-groups.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using_cluster_computing.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribePlacementGroups.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreatePlacementGroup.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DeletePlacementGroup.html}
}
