% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instances.r
\name{reboot_instances}
\alias{reboot_instances}
\title{Reboot EC2 Instance}
\usage{
reboot_instances(instance, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Reboot a running EC2 instance
}
\examples{
\dontrun{
i <- run_instance()
reboot_instances(i[[1]])
stop_instances(i[[1]])
terminate_instances(i[[1]])
}
}

