% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecurityGroups.r
\name{describe_sgroups}
\alias{create_sgroup}
\alias{delete_sgroup}
\alias{describe_sgroups}
\title{Security Groups}
\usage{
describe_sgroups(id, name, filter, ...)

create_sgroup(name, description, vpc, ...)

delete_sgroup(name, id, ...)
}
\arguments{
\item{id}{A character string specifying a security group ID.}

\item{name}{A character string (max 255 characters) specifying a security group name.}

\item{filter}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{description}{A character string specifying a security group description.}

\item{vpc}{A character string specifying a VPC Id (required for a VPC).}
}
\value{
For \code{describe_sgroups} and \code{create_sgroup}, a list of objects of class \dQuote{ec2_security_group}. For \code{delete_sgroup}, a logical.
}
\description{
Describe, create, and delete Security Groups
}
\examples{
\dontrun{
describe_sgroups()
sg <- create_sgroup("test_group", "example security group")
delete_sgroup(sg)
}
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-network-security.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeSecurityGroups.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreateSecurityGroup.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DeleteSecurityGroup.html}
}

