% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instances.r
\name{start_instances}
\alias{start_instances}
\alias{stop_instances}
\title{Stop/Start EC2 Instance}
\usage{
start_instances(instance, info, ...)

stop_instances(instance, force, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}.}

\item{info}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{force}{\dots}
}
\value{
A list
}
\description{
Stop a running or start a stopped EC2 instance
}
\examples{
\dontrun{
i <- run_instance()
stop_instances(i)
start_instances(i)
stop_instances(i)
terminate_instances(i)
}
}

