% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subnets.r
\name{describe_subnets}
\alias{create_subnet}
\alias{delete_subnet}
\alias{describe_subnets}
\title{Subnets}
\usage{
describe_subnets(subnet, filter, ...)

create_subnet(vpc, cidr, zone, ...)

delete_subnet(subnet, ...)
}
\arguments{
\item{subnet}{A character string containing the name of a subnet, or an object of class \dQuote{ec2_subnet}.}

\item{filter}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Get, create, and delete subnets
}
\examples{
\dontrun{
# RStudio AMIs from: http://www.louisaslett.com/RStudio_AMI/
get_image_attr("ami-7f9dc615", "description")
}
}

