% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tags.r
\name{create_tags}
\alias{create_tags}
\alias{delete_tags}
\alias{describe_tags}
\title{Tags}
\usage{
create_tags(resource, tag, ...)

delete_tags(resource, tag, ...)

describe_tags(filter, n, page, ...)
}
\arguments{
\item{resource}{A character vector specifying one or more resource IDs, typically EC2 instance IDs.}

\item{tag}{A named character string of key-value pairs of tag names and their corresponding values.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{filter}{\dots}

\item{n}{\dots}

\item{page}{\dots}
}
\value{
A logical.
}
\description{
Tag an EC2 Resource
}
\seealso{
\code{\link{associate_ip}}
}

