% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instances.r
\name{terminate_instances}
\alias{terminate_instances}
\title{Terminate EC2 Instance}
\usage{
terminate_instances(instance, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Terminate a stopped EC2 instance
}
\examples{
\dontrun{
i <- run_instance()
stop_instances(i[[1]])
terminate_instances(i[[1]])
}
}

