% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.R
\name{create_vpc}
\alias{create_vpc}
\alias{delete_vpc}
\title{create_vpc}
\usage{
create_vpc(cidr, tenancy = c("default", "dedicated"), ...)

delete_vpc(vpc, ...)
}
\arguments{
\item{cidr}{A character string specifying a network range for the subnet in CIDR notation.}

\item{tenancy}{A character string specifying either \dQuote{default} or \dQuote{dedicated} (single-tenant) hardware for instances launched within the VPC.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{vpc}{A character string specifying a VPC Allocation ID or an object of class \dQuote{ec2_ip}.}
}
\value{
For \code{create_vpc}, a list of class \dQuote{ec2_vpc}. For \code{delete_vpc}, a logical indicating whether the operation succeeded.
}
\description{
Create/delete VPCs
}
\examples{
\dontrun{
v <- create_vpc(cidr = "10.0.0.0/16")
describe_vpcs()

# delete VPC
delete_vpc(v)
}
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreateVpc.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DeleteVpc.html}
}
\seealso{
\code{\link{describe_vpcs}}, \code{\link{get_vpc_attr}}
}
\keyword{security}
