% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.r
\name{allocate_ip}
\alias{allocate_ip}
\alias{release_ip}
\title{Allocate/Release IP Address}
\usage{
allocate_ip(domain, ...)

release_ip(allocation, public, ...)
}
\arguments{
\item{domain}{Optionally, a character string specifying \dQuote{vpc} or \dQuote{standard}.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{allocation}{A character string specifying an EC2 IP address Allocation ID.}

\item{public}{This is the public IP address for a \dQuote{classic} EC2 instance. For VPC, use \code{allocation} instead.}
}
\value{
For \code{allocate_ip}, a list containing the IP address. For \code{release_ip}, a logical.
}
\description{
Allocate or release VPC or standard IP Address
}
\seealso{
\code{\link{associate_ip}}
}

