% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.R
\name{associate_ip}
\alias{associate_ip}
\alias{disassociate_ip}
\title{(Dis)Associate IP}
\usage{
associate_ip(instance, ip, private, netinterface, allow, ...)

disassociate_ip(ip, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}. Or, actions can be taken on multiple instances in one call by passing a character vector of instance IDs, or a list of objects of class \dQuote{ec2_instance}.}

\item{ip}{An object of class \dQuote{ec2_ip} or an allocationId (for a VPC IP) or publicIp (for a standard IP).}

\item{private}{For a VPC \code{ip}, the primary or secondary private IP address to associate with the Elastic IP address. If no private IP address is specified, the Elastic IP address is associated with the primary private IP address.}

\item{netinterface}{For a VCP \code{ip}, the ID of the network interface. If the instance has more than one network interface, you must specify a network interface ID.}

\item{allow}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list.
}
\description{
Associate/Disassociate IP with Instance
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/elastic-ip-addresses-eip.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_AssociateAddress.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DisassociateAddress.html}
}
\seealso{
\code{\link{allocate_ip}}, \code{\link{describe_ips}}, \code{\link{release_ip}}
}
