\name{aws.ec2-package}
\alias{aws.ec2-package}
\alias{aws.ec2}
\docType{package}
\title{AWS EC2 Client Package}
\description{A simple client package for the Amazon Web Services (AWS) Elastic Cloud Compute (EC2) REST API.}
\details{}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
