% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.ec2-package.R
\docType{package}
\name{aws.ec2-package}
\alias{aws.ec2-package}
\alias{aws.ec2}
\title{AWS EC2 Client Package}
\description{
A simple client package for the Amazon Web Services (AWS) Elastic Cloud Compute (EC2) REST API.
}
\seealso{
\code{\link[=account_attrs]{account_attrs()}} \code{\link[=run_instances]{run_instances()}}
}
\author{
Thomas J. Leeper \href{mailto:thosjleeper@gmail.com}{thosjleeper@gmail.com}
}
\keyword{package}
