% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Images.R
\name{describe_images}
\alias{describe_images}
\title{Describe AMI(s)}
\usage{
describe_images(image, filter, availableto, owner, ...)
}
\arguments{
\item{image}{A character string containing an Amazon Machine Image (AMI) image ID or an object of class \dQuote{ec2_image}.}

\item{filter}{\dots}

\item{availableto}{\dots}

\item{owner}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Search/Describe AMI(s)
}
\examples{
\dontrun{
# RStudio AMIs from: http://www.louisaslett.com/RStudio_AMI/
describe_images("ami-7f9dc615")

# Amazon Linux AMI from: http://aws.amazon.com/amazon-linux-ami/
describe_images("ami-08111162")
}
}
\keyword{images}
