% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.r
\name{describe_ip}
\alias{describe_ip}
\title{Describe IP}
\usage{
describe_ip(allocation, public, filter, ...)
}
\arguments{
\item{allocation}{A character string specifying an EC2 IP address Allocation ID.}

\item{public}{\dots}

\item{filter}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Get IP information
}

