% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebs_snapshots.R
\name{create_snapshot}
\alias{create_snapshot}
\alias{delete_snapshot}
\title{Elastic Block Store Snapshots}
\usage{
create_snapshot(volume, description = NULL, ...)

delete_snapshot(snapshot, ...)
}
\arguments{
\item{volume}{An object of class \dQuote{ec2_ebs_volume} or a character string containing a volume ID.}

\item{description}{\dots}

\item{\dots}{\dots}

\item{snapshot}{A character string containing an EBS Volume Snapshot snapshot ID or an object of class \dQuote{ec2_ebs_snapshot}.}

\item{volume}{A character string containing an EBS Volume volume ID, or an object of class \dQuote{ec2_ebs_volume}.}
}
\value{
\dots
}
\description{
Manage Elastic Block Store (EBS) volumes for EC2
}
\references{
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreateVolume.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DeleteVolume.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_ModifyVolume.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_AttachVolume.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DetachVolume.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVolumes.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVolumeStatus.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVolumeAttribute.html}
\url{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_ModifyVolumeAttribute.html}
}
\seealso{
\code{\link[=run_instances]{run_instances()}}, \code{\link[=create_volume]{create_volume()}}
}
