% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{ec2HTTP}
\alias{ec2HTTP}
\title{EC2 API Requests}
\usage{
ec2HTTP(query = list(), dryrun, region = Sys.getenv("AWS_DEFAULT_REGION",
  "us-east-1"), key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"), version = "2015-10-01",
  clean_response = TRUE, ...)
}
\arguments{
\item{query}{A named list of query string parameters.}

\item{dryrun}{An optional logical specifying whether to execute a consequence-free \dQuote{dry run} of the request.}

\item{region}{A character string containing the AWS region. If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an AWS Access Key ID. If missing, defaults to value stored in environment variable \dQuote{AWS_ACCESS_KEY_ID}.}

\item{secret}{A character string containing an AWS Secret Access Key.  If missing, defaults to value stored in environment variable \dQuote{AWS_SECRET_ACCESS_KEY}.}

\item{version}{A character string specifying an API version. Default is \dQuote{2015-10-01}.}

\item{clean_response}{A logical indicating whether to remove line breaks from the response. This is useful for returning a clean response object, but may not be appropriate if the XML-formatted API response contains meaningful linebreaks (e.g., in a keypair).}

\item{...}{Additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
A list
}
\description{
Execute an EC2 API Request
}
