% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gateways.R
\name{attach_gateway}
\alias{attach_gateway}
\alias{detach_gateway}
\title{Attach/detach Gateways}
\usage{
attach_gateway(gateway, vpc, ...)

detach_gateway(gateway, vpc, ...)
}
\arguments{
\item{gateway}{A character string specifying an Internet Gateway ID, or an object of class \dQuote{ec2_internet_gateway}.}

\item{vpc}{A character string specifying a VPC Allocation ID or an object of class \dQuote{ec2_ip}.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\description{
Attach or Detacth Network Gateways for a VPC
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_AttachInternetGateway.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DetachInternetGateway.html}
}
\seealso{
\code{\link{create_gateway}}, \code{\link{describe_gateways}}
}
\keyword{security}
