% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_console_output.R
\name{get_console_output}
\alias{get_console_output}
\title{Get EC2 Instance Console Output}
\usage{
get_console_output(instance, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}. Or, actions can be taken on multiple instances in one call by passing a character vector of instance IDs, or a list of objects of class \dQuote{ec2_instance}.}

\item{...}{Additional arguments passed to [ec2HTTP()].}
}
\value{
A character string
}
\description{
Retrieve console output for an EC2 instance as a character string
}
\examples{
\dontrun{
i <- run_instances()
cat(get_console_output(), "\n")
stop_instances(i[[1]])
terminate_instances(i[[1]])
}
}
\references{
<http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_GetConsoleOutput.html>
}
\seealso{
[get_instance_attr()], [describe_instances()], [get_password_data()]
}
\keyword{instances}
