% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instance_attr.R
\name{get_console_output}
\alias{get_console_output}
\title{Get EC2 Instance Console Output}
\usage{
get_console_output(instance, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}. Or, actions can be taken on multiple instances in one call by passing a character vector of instance IDs, or a list of objects of class \dQuote{ec2_instance}.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A character string
}
\description{
Retrieve console output for an EC2 instance as a character string
}
\examples{
\dontrun{
i <- run_instances()
cat(get_console_output(), "\\n")
stop_instances(i[[1]])
terminate_instances(i[[1]])
}
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_GetConsoleOutput.html}
}
\seealso{
\code{\link{get_instance_attr}}, \code{\link{describe_instances}}, \code{\link{get_password_data}}
}
\keyword{instances}
