% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instance_attr.R
\name{get_password_data}
\alias{get_password_data}
\title{Get EC2 Instance Password Data}
\usage{
get_password_data(instance, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}. Or, actions can be taken on multiple instances in one call by passing a character vector of instance IDs, or a list of objects of class \dQuote{ec2_instance}.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Retrieve password data for an EC2 instance
}
\examples{
\dontrun{
i <- run_instances()
get_password_data(i[[1]])
stop_instances(i[[1]])
terminate_instances(i[[1]])
}
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_GetConsoleOutput.html}
}
\seealso{
\code{\link{describe_instances}}, \code{\link{get_console_output}}
}
\keyword{instances}
