% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instances.r
\name{monitor_instances}
\alias{monitor_instances}
\alias{unmonitor_instances}
\title{EC2 Instance Monitoring}
\usage{
monitor_instances(instance, ...)

unmonitor_instances(instance, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Set EC2 instance monitoring on or off
}
\examples{
\dontrun{
i <- run_instance()
monitor_instances(i)
unmonitor_instances(i)

stop_instances(i)
terminate_instances(i)
}
}

