% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instances.r
\name{run_instances}
\alias{run_instances}
\title{Run an EC2 Instance}
\usage{
run_instances(image, type, min = 1, max = min, keypair, subnet, userdata,
  ...)
}
\arguments{
\item{image}{A character string containing an Amazon Machine Image (AMI) image ID or an object of class \dQuote{ec2_image}.}

\item{type}{A character string specifying the type of EC2 instance to use. See \url{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-types.html} for details of types and available options.}

\item{min}{An integer specifying a minimum number of instances to launch. Defaults to 1.}

\item{keypair}{A character string (max 255 character) containing the name of a keypair, or an object of class \dQuote{ec2_keypair}.}

\item{subnet}{A character string containing the name of a subnet, or an object of class \dQuote{ec2_subnet}.}

\item{userdata}{Optionally, a character string specifying a script to run during launch.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{min}{An integer specifying a minimum number of instances to launch. Defaults to \code{min}.}
}
\value{
A list
}
\description{
Run/launch a new EC2 Instance
}
\examples{
\dontrun{
# RStudio AMIs from: http://www.louisaslett.com/RStudio_AMI/
describe_images("ami-7f9dc615")
s <- describe_subnets()
i <- run_instances("ami-7f9dc615", type = "t2.micro", subnet = s[[1]])

stop_instances(i[[1]])
terminate_instances(i[[1]])
}
}

