% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecurityGroups.r
\name{describe_sgroups}
\alias{create_sgroup}
\alias{delete_sgroup}
\alias{describe_sgroups}
\title{Security Groups}
\usage{
describe_sgroups(name, id, filter, ...)

create_sgroup(name, description, vpc, ...)

delete_sgroup(name, id, ...)
}
\arguments{
\item{name}{A character string (max 255 characters) specifying a security group name.}

\item{id}{A character string specifying a security group ID.}

\item{filter}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{description}{A character string specifying a security group description.}

\item{vpc}{A character string specifying a VPC Id (required for a VPC).}
}
\value{
For \code{describe_sgroups}, a list of objects of class \dQuote{ec2_security_group}.
}
\description{
Describe, create, and delete Security Groups
}

