% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecurityGroups_rules.R, R/ip.R
\name{authorize_ingress}
\alias{authorize_ingress}
\alias{revoke_ingress}
\alias{my_ip}
\title{Security Group Ingress}
\usage{
authorize_ingress(sgroup, cidr = paste0(my_ip(), "/32"), port,
  protocol = "-1", ...)

revoke_ingress(sgroup, cidr = paste0(my_ip(), "/32"), port, protocol = "-1",
  ...)

my_ip()
}
\arguments{
\item{sgroup}{A character vector specifying one or more Security Group IDs, or (a list of) objects of class \dQuote{ec2_security_group}.}

\item{cidr}{A one- or two-element character vector specifying an IP or range of IP addresses. The default is your current machine's public IP as returned by \url{http://checkip.amazonaws.com/}.}

\item{port}{A one- or two-element integer vector, specifying a port or port range.}

\item{protocol}{A character string specifying a protocol. A value of \dQuote{-1} (the default) means all protocols.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}
}
\value{
A list
}
\description{
Authorize/Revoke Security Group Ingress
}
\examples{
\dontrun{
s <- describe_subnets()
g <- create_sgroup("test_group", "new security group", vpc = s[[1]])
authorize_ingress(g, port = 80, protocol = "tcp")

# cleanup
revoke_ingress(g, port = 80, protocol = "tcp")
}
}
\seealso{
\code{\link{create_sgroup}}
}
\keyword{security}
