% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instances.R
\name{start_instances}
\alias{start_instances}
\alias{stop_instances}
\alias{terminate_instances}
\alias{reboot_instances}
\title{Start, Stop, Reboot, or Terminate EC2 Instance}
\usage{
start_instances(instance, info, ...)

stop_instances(instance, force, ...)

terminate_instances(instance, ...)

reboot_instances(instance, ...)
}
\arguments{
\item{instance}{A character string containing an EC2 instance ID or an object of class \dQuote{ec2_instance}. Or, actions can be taken on multiple instances in one call by passing a character vector of instance IDs, or a list of objects of class \dQuote{ec2_instance}.}

\item{info}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{force}{\dots}
}
\value{
A list
}
\description{
Stop a running instance, terminate a stopped instance, start a stopped instance, or reboot an instance.
}
\examples{
\dontrun{
i <- run_instances()
stop_instances(i[[1]])
start_instances(i[[1]])
stop_instances(i[[1]])
terminate_instances(i[[1]])
}
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_StartInstances.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_StopInstances.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_TerminateInstances.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_RebootInstances.html}
}
\seealso{
\code{\link{describe_instances}}, \code{\link{run_instances}}
}
\keyword{instances}
