% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subnets.R
\name{describe_subnets}
\alias{describe_subnets}
\alias{create_subnet}
\alias{delete_subnet}
\title{Subnets}
\usage{
describe_subnets(subnet, filter, ...)

create_subnet(vpc, cidr, zone, ...)

delete_subnet(subnet, ...)
}
\arguments{
\item{subnet}{A character string containing the name of a subnet, or an object of class \dQuote{ec2_subnet}.}

\item{filter}{\dots}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{vpc}{A character string specifying a VPC Allocation ID or an object of class \dQuote{ec2_ip}.}

\item{cidr}{A character string specifying a network range for the subnet in CIDR notation.}

\item{zone}{Optionally, a character string specifying an availability zone (see \code{\link{describe_zones}}) or an object of class \dQuote{ec2_zone}. If omitted, a zone is selected automatically.}
}
\value{
A list of objects of class \dQuote{ec2_subnet}.
}
\description{
Get, create, and delete subnets
}
\details{
Use \code{\link{create_netacl}} and \code{\link{associate_netacl}} to specify a non-default Access Control List (ACL) for the subnet
}
\examples{
\dontrun{
# describe existing subnets
(d <- describe_subnets())
describe_subnets(d[[1]])

# create a subnet
## setup new VPC IP address
new_ip <- allocate_ip("vpc")
aid <- describe_ips(new_ip$publicIp)[[1]]$allocationId
s <- create_subnet(aid, cidr = "10.0.1.0/24")

# delete a subnet

}
}
\references{
\url{http://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_Subnets.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeSubnets.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreateSubnet.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DeleteSubnet.html}
}
\seealso{
\code{\link{create_vpc}}, \code{\link{create_netacl}}, \code{\link{associate_netacl}}
}
\keyword{security}
