% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_tags.R
\name{create_tags}
\alias{create_tags}
\alias{delete_tags}
\alias{describe_tags}
\title{Tags}
\usage{
create_tags(resource, tag, ...)

delete_tags(resource, tag = NULL, ...)

describe_tags(filter = NULL, n = NULL, page = NULL, ...)
}
\arguments{
\item{resource}{A character vector specifying one or more resource IDs, typically EC2 instance IDs.}

\item{tag}{A named character string of key-value pairs of tag names and their corresponding values. For `delete_tags`, the value can be an empty string (in which case, the tag is delete regardless of value) or a specific value (in which case the tag is only deleted if it matches the value).}

\item{...}{Additional arguments passed to [ec2HTTP()].}

\item{filter}{\dots}

\item{n}{\dots}

\item{page}{\dots}
}
\value{
For `create_tags` and `delete_tags`, a logical. Otherwise, a list of tags.
}
\description{
Tag an EC2 Resource
}
\examples{
\dontrun{
create_tags("i-b79cfd34", list(foo = "bar"))
describe_tags()
delete_tags("i-b79cfd34", list("foo" = "notbar"))
delete_tags("i-b79cfd34", list("foo" = "bar"))
}
}
\references{
<http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_Tags.html>
<http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeTags.html>
<http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreateTags.html>
<http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeTags.html>
}
\seealso{
[describe_instances()], [get_instance_attr()], [associate_ip()]
}
\keyword{instances}
