% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tags.R
\name{create_tags}
\alias{create_tags}
\alias{delete_tags}
\alias{describe_tags}
\title{Tags}
\usage{
create_tags(resource, tag, ...)

delete_tags(resource, tag, ...)

describe_tags(filter, n, page, ...)
}
\arguments{
\item{resource}{A character vector specifying one or more resource IDs, typically EC2 instance IDs.}

\item{tag}{A named character string of key-value pairs of tag names and their corresponding values. For \code{delete_tags}, the value can be an empty string (in which case, the tag is delete regardless of value) or a specific value (in which case the tag is only deleted if it matches the value).}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{filter}{\dots}

\item{n}{\dots}

\item{page}{\dots}
}
\value{
For \code{create_tags} and \code{delete_tags}, a logical. Otherwise, a list of tags.
}
\description{
Tag an EC2 Resource
}
\examples{
\dontrun{
create_tags("i-b79cfd34", list(foo = "bar"))
describe_tags()
delete_tags("i-b79cfd34", list("foo" = "notbar"))
delete_tags("i-b79cfd34", list("foo" = "bar"))
}
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_Tags.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeTags.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreateTags.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeTags.html}
}
\seealso{
\code{\link{describe_instances}}, \code{\link{get_instance_attr}}, \code{\link{associate_ip}}
}
\keyword{instances}
