% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.R
\name{set_vpc_attr}
\alias{set_vpc_attr}
\alias{get_vpc_attr}
\title{VPC Attributes}
\usage{
set_vpc_attr(vpc, dns, hostnames, ...)

get_vpc_attr(vpc, attribute = c("enableDnsSupport", "enableDnsHostnames"),
  ...)
}
\arguments{
\item{vpc}{A character string specifying a VPC Allocation ID or an object of class \dQuote{ec2_ip}.}

\item{dns}{A logical indicating whether DNS resolution is supported for the VPC. A request can only specify only \code{dns} xor \code{hostnames}.}

\item{hostnames}{A logical indicating whether instances launched in the VPC get DNS hostnames. A request can only specify only \code{dns} xor \code{hostnames}.}

\item{...}{Additional arguments passed to \code{\link{ec2HTTP}}.}

\item{attribute}{A character string specifying one of \dQuote{enableDnsSupport} or \dQuote{enableDnsHostnames}.}
}
\value{
For \code{get_vpc_attr}, a logical indicating the value of the attribute. For \code{set_vpc_attr}, a logical indicating whether the operation succeeded.
}
\description{
Get and set VPC attributes
}
\examples{
\dontrun{
# create VPC
v <- create_vpc("10.0.0.0/16")
get_vpc_attr(v, "enableDnsSupport")
set_vpc_attr(v, dns = FALSE)

# cleanup
delete_vpc(v)
}
}
\references{
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVpcAttribute.html}
\url{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_ModifyVpcAttribute.html}
}
\seealso{
\code{\link{describe_vpcs}}, \code{\link{create_vpc}}
}
\keyword{security}
